requirements.txt
altgraph==0.17.4
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
packaging==24.1
pefile==2023.2.7
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
urllib3==2.2.2
